examples/content-recommender/ai-pipeline/requirements.txt
# Base dependencies (from blissful-infra ai-pipeline plugin)
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
scikit-learn==1.5.0
numpy==1.26.4
kafka-python-ng==2.2.2
mlflow-skinny==2.13.0
clickhouse-connect==0.7.19

# Recommendation-specific
implicit==0.7.2        # Fast ALS collaborative filtering (CPU)
scipy==1.13.0          # Sparse matrix operations for user-item matrix
